requirements.txt
future~=0.18.3
ttkbootstrap~=1.10.1
Pillow~=10.0.0
chardet~=5.2.0
lxml~=4.9.3